requirements.txt
fastapi==0.118.0
uvicorn[standard]==0.37.0
jinja2==3.1.6
python-multipart==0.0.20
livekit-agents[google,images]~=1.0,>=1.0.18
livekit-plugins-noise-cancellation~=0.2
python-dotenv
asyncio
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1
psycopg2-binary==2.9.10
pandas==2.1.4
openpyxl==3.1.2